UserManagement CRUD Ok with unauthorized exception

+++ Library-api/library-api-business/src/main/java/org/alain/library/api/business/impl/CrudManagerImpl.java
@@ -26,7 +26,8 @@ public class CrudManagerImpl<T> implements CrudManager<T> {
 
     @Override
     public void delete(Long id){
-        repository.deleteById(id);
+        if(repository.findById(id).isPresent())
+            repository.deleteById(id);
     }
 
     @Override
